requirements.txt
supyr-struct
selection
asyncio
pybase64
sockets
threaded
websockets
pyyaml